refactored to traits
